Combat improvements.
Added vfx to show when you hit on beat and off beat. Also added vfx to show if you block on beat or off beat. Made it easier to hit enemies with your weapon. Made it so that if an enemy is just barely out of reach for your weapon then your character will close the remaining distance.